+++ src/libs/tmc/TmcVerifier/verifier.cpp
@@ -64,7 +64,7 @@ void TmcVerifier::setRealTypeEnabled(bool isRealTypeEnabled)
     m_converter->setRealTypeEnabled(isRealTypeEnabled);
 }
 
-void TmcVerifier::setDelta(std::optional<int> value)
+void TmcVerifier::setDelta(std::optional<float> value)
 {
     m_converter->setDelta(std::move(value));
 }
